+++ scripts/createDmg.sh
@@ -6,9 +6,18 @@
 # Created by Derek Clarkson on 27/08/10.
 # Copyright 2010 Derek Clarkson. All rights reserved.
 
-echo "Copying docset ..."
-mv -vf "$DC_BUILD_DIR/appledoc/docset" "$DC_ARTIFACT_DIR/au.com.dhc.$DC_PROJECT_NAME.docset"
-cp installDocSet "$DC_ARTIFACT_DIR"
+DOCSET_PATH="/Users/derek/Library/Developer/Shared/Documentation/DocSets/$DC_COMPANY_ID.$DC_PROJECT_NAME.docset"
+echo "Checking for docset at $DOCSET_PATH"
+if [ -d $DOCSET_PATH ]; then
+	echo "Copying docset ..."
+	cp -fR "$DOCSET_PATH" "$DC_ARTIFACT_DIR"
+	
+	echo "Writing install docset script file ..."
+	echo "echo \"Installing $DC_PROJECT_NAME documentation into XCode.\"" > "$DC_ARTIFACT_DIR/installDocSet"
+	echo "osascript -e 'tell application \"Xcode\" to load documentation set with path (system attribute \"PWD\") & \"/$DC_COMPANY_ID.$DC_PROJECT_NAME.docset\"'" >> "$DC_ARTIFACT_DIR/installDocSet"
+	chmod a+x "$DC_ARTIFACT_DIR/installDocSet"
+
+fi
 
 echo "Copying markdown documentation from root of project ..."
 cp -v *.markdown "$DC_ARTIFACT_DIR"